Significant changes to structure.
Renamed a couple of directories and shuffled things around attempting a slightly clearer naming scheme for the directories.

Of most importance, changed the 'public' folder to 'client/web' and added several stubs for mobile clients, and changes the 'routes' folder to 'api'.

Added quite a bit of information to the tutorial about the structure and made several very small code changes as a result of the new structure.